move .css locations and touch ups
