+++ src/components/Auth/index.tsx
@@ -1,6 +1,5 @@
 
 import { useUser } from '@auth0/nextjs-auth0';
-import { setSyntheticTrailingComments } from 'typescript';
 
 
 const AuthComponent = () => {
