+++ src/components/HeaderBar/HeaderBar.module.scss
@@ -1,0 +1,61 @@
+.Content {
+  background-color: #ffffff;
+  height: 80px;
+  display: flex;
+  margin: 0;
+}
+
+@media screen and (max-width: 1000px) {
+  .Content {
+    height: 60px;
+    align-items: center;
+    justify-content: center;
+    width: 100%;
+    padding-left: 10px;
+  }
+  
+  .LogoContainer {
+    width: 10%;  
+    height: 20px;
+    .Logo {
+      object-fit: contain;
+      width: 100% !important;
+      position: relative !important;
+      height: unset !important;
+    }
+  }
+  .SearchContainer {
+    padding: 5px;
+    height: 30px;
+    background-color: #E3E3E3;
+    width: 60%;
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    justify-content: center;
+    border-radius: 2px;
+    margin-left: 15px;
+    margin-right: 7px;
+  }
+  .TKlogo {
+    font-size: 4rem;
+  }
+  .SearchBox {
+    width: 70%;
+    margin: 8px;
+    height: 20px;
+    border: none;
+    background: none;
+    font-size: 1em;
+  }
+  .Profile {
+    height: 28px;
+    width: 28px;
+    background-color: #039941;
+    border-radius: 100%;
+    margin: 8px;
+  }
+  .menu-bars {
+    margin: 5px;
+  }
+}
