+++ src/components/Map/Map.module.css
@@ -1,0 +1,8 @@
+.Map {
+    display: flex;
+    flex-direction: horizontal;
+    justify-content: center;
+    justify-self: center;
+    height: 100vh;
+    width: 100vw;
+  }
